+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Bug-070173 - shopify data migration not worrking as tag lang table not exisit in current release
 - Bug-070827 - Adding special price to batch products in google shopping feed.
 - Bug-070881 - Easypost : Unable to validate plugin API keys
+- Bug-070882 - Unable to save lang data of Tax category, Add Brand, Add Category from Admin > Add Product.
 # RV-10.0.0.20221123
 
 ## Bug Fixes: